refactor: localize storage core compat consumers
